refactor: optimize.optimize should never get a BytesIO object
the only time we use the optimize function is in application.py,
after checking that we're not sending it a BytesIO object.

allowing the optimize function to accept and return complicates
subsequent logic, which assumes that the returned value has image-specific
methods like format and save.

simpler to disallow sending svgs to the optimize function, which
has no business getting them anyway.